The code is there and it compiles but breaks when you try to add any entity for reasons unknown.
The static addSphere, addBlock, and addRagdoll methods were changed to non-static to work with the new stack (necessitating other edits).

As a test I changed the control for addSphere to instead removeLast, as I didn't want to try to add to the UI.

The new classes as EntityStack and StackNode.  They have yet to be properly unit tested, but they are relatively simple.